+++ AuthPermissions/DataLayer/Migrations/AuthPermissionsDbContextModelSnapshot.cs
@@ -156,17 +156,12 @@ namespace AuthPermissions.DataLayer.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<string>("AuthUserUserId")
-                        .HasColumnType("nvarchar(256)");
-
                     b.Property<byte[]>("ConcurrencyToken")
                         .IsConcurrencyToken()
                         .ValueGeneratedOnAddOrUpdate()
                         .HasColumnType("ROWVERSION");
 
                     b.HasKey("UserId", "RoleName");
-
-                    b.HasIndex("AuthUserUserId");
 
                     b.HasIndex("RoleName");
 
@@ -193,13 +188,15 @@ namespace AuthPermissions.DataLayer.Migrations
 
             modelBuilder.Entity("AuthPermissions.DataLayer.Classes.UserToRole", b =>
                 {
-                    b.HasOne("AuthPermissions.DataLayer.Classes.AuthUser", null)
-                        .WithMany("UserRoles")
-                        .HasForeignKey("AuthUserUserId");
-
                     b.HasOne("AuthPermissions.DataLayer.Classes.RoleToPermissions", "Role")
                         .WithMany()
                         .HasForeignKey("RoleName")
+                        .OnDelete(DeleteBehavior.Cascade)
+                        .IsRequired();
+
+                    b.HasOne("AuthPermissions.DataLayer.Classes.AuthUser", null)
+                        .WithMany("UserRoles")
+                        .HasForeignKey("UserId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();